+++ keystorage/hashicorp/src/testFixtures/java/tech/pegasys/signing/hashicorp/dsl/certificates/SelfSignedCertificate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 ConsenSys AG.
+ * Copyright ConsenSys AG.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -9,6 +9,8 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
  */
 package tech.pegasys.signing.hashicorp.dsl.certificates;
 
@@ -22,9 +24,7 @@ import java.nio.file.Path;
 import java.security.GeneralSecurityException;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
-import java.security.PrivateKey;
 import java.security.SecureRandom;
-import java.security.cert.Certificate;
 import java.security.cert.CertificateEncodingException;
 import java.security.cert.X509Certificate;
 import java.time.Instant;
@@ -34,7 +34,6 @@ import java.util.Date;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
-
 import org.bouncycastle.asn1.x500.X500Name;
 import org.bouncycastle.asn1.x509.BasicConstraints;
 import org.bouncycastle.asn1.x509.Extension;
@@ -51,36 +50,37 @@ import org.bouncycastle.operator.jcajce.JcaContentSignerBuilder;
 import org.bouncycastle.util.io.pem.PemObject;
 import org.bouncycastle.util.io.pem.PemWriter;
 
-/**
- * Utility class to generate temporary self-signed certificates with SAN extension for multiple DNS
- * and IP addresses in PEM formats for testing purposes using BouncyCastle APIs.
- *
- * <p>Note: DO NOT USE IN PRODUCTION. The generated stores and files are marked to be deleted on JVM
- * exit.
- */
-public final class SelfSignedCertificate {
+public class SelfSignedCertificate {
+
   private static final BouncyCastleProvider BOUNCY_CASTLE_PROVIDER = new BouncyCastleProvider();
   private static final boolean IS_CA = true;
   private static final String distinguishedName = "CN=localhost";
   private static final List<String> sanHostNames = List.of("localhost");
   private static final List<String> sanIpAddresses = List.of("127.0.0.1");
-  private final Path certificatePath;
-  private final Path privateKeyPath;
 
-  private SelfSignedCertificate(final Path certificatePath, final Path privateKeyPath) {
-    this.certificatePath = certificatePath;
-    this.privateKeyPath = privateKeyPath;
+  private final KeyPair keyPair; // not sure if this is needed.
+  private final X509Certificate certificate;
+
+  public SelfSignedCertificate(final KeyPair keyPair, final X509Certificate certificate) {
+    this.keyPair = keyPair;
+    this.certificate = certificate;
   }
 
   public static SelfSignedCertificate generate() {
     try {
       final KeyPair keyPair = generateKeyPair();
       return new SelfSignedCertificate(
-          createPemCertificateFile(generateSelfSignedCertificate(keyPair)),
-          createPemPrivateKeyFile(keyPair.getPrivate()));
+          keyPair,
+          generateSelfSignedCertificate(keyPair));
     } catch (final Exception e) {
       throw new RuntimeException(e);
     }
+  }
+
+  private static KeyPair generateKeyPair() throws GeneralSecurityException {
+    final KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
+    keyPairGenerator.initialize(2048, new SecureRandom());
+    return keyPairGenerator.generateKeyPair();
   }
 
   private static X509Certificate generateSelfSignedCertificate(final KeyPair keyPair)
@@ -111,12 +111,6 @@ public final class SelfSignedCertificate {
         .getCertificate(v3CertificateBuilder.build(contentSigner));
   }
 
-  private static KeyPair generateKeyPair() throws GeneralSecurityException {
-    final KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance("RSA");
-    keyPairGenerator.initialize(2048, new SecureRandom());
-    return keyPairGenerator.generateKeyPair();
-  }
-
   private static GeneralNames getSubjectAlternativeNames() {
     final List<GeneralName> hostGeneralNames =
         sanHostNames.stream()
@@ -134,35 +128,22 @@ public final class SelfSignedCertificate {
     return new GeneralNames(generalNames);
   }
 
-  private static Path createPemCertificateFile(final Certificate certificate)
-      throws IOException, CertificateEncodingException {
-    final Path certificatePath = Files.createTempFile("test", ".crt");
-    certificatePath.toFile().deleteOnExit();
+  public X509Certificate getCertificate() {
+    return certificate;
+  }
 
-    try (final BufferedWriter writer = Files.newBufferedWriter(certificatePath, UTF_8);
+  public void writePrivateKeyToFile(final Path outputFile) throws IOException {
+    try (final BufferedWriter writer = Files.newBufferedWriter(outputFile, UTF_8);
+        final PemWriter pemWriter = new PemWriter(writer)) {
+      pemWriter.writeObject(new PemObject("PRIVATE KEY", keyPair.getPrivate().getEncoded()));
+    }
+  }
+
+  public void writeCertificateToFile(final Path pemFile)
+      throws IOException, CertificateEncodingException {
+    try (final BufferedWriter writer = Files.newBufferedWriter(pemFile, UTF_8);
         final PemWriter pemWriter = new PemWriter(writer)) {
       pemWriter.writeObject(new PemObject("CERTIFICATE", certificate.getEncoded()));
     }
-
-    return certificatePath;
-  }
-
-  private static Path createPemPrivateKeyFile(final PrivateKey privateKey) throws IOException {
-    final Path privateKeyPath = Files.createTempFile("test", ".key");
-    privateKeyPath.toFile().deleteOnExit();
-
-    try (final BufferedWriter writer = Files.newBufferedWriter(privateKeyPath, UTF_8);
-        final PemWriter pemWriter = new PemWriter(writer)) {
-      pemWriter.writeObject(new PemObject("PRIVATE KEY", privateKey.getEncoded()));
-    }
-    return privateKeyPath;
-  }
-
-  public Path certificatePath() {
-    return certificatePath;
-  }
-
-  public Path privateKeyPath() {
-    return privateKeyPath;
   }
 }
